portpercost

+++ 3_Implementation/src/earning.c
@@ -1,6 +1,6 @@
 #include<header.h>
 
-float earning(int consumed_unit, int rate, int time_consumed)
+float earning(int consumed_unit, int rate, float time_consumed)
 {
     float earn= consumed_unit*rate;
     return earn;
